updated identifer and reference data elements. added MessageHeader.response
Assumed that the id data element for the requested Task instance id was the same as requested. E.g., in the initialSubmissionBySponser id=bd675726-6ca7-4fce-8edb-6f29c0df2bcc is requested by Sponsor. This is not guaranteed in real-life, and there are many alternate Task workflows. However, the reference data type in FHIR contains an identifier data element. I populated this with the Task identifier.   Recall that the FHIR data type has a system and value. A recipient can use this in a couple of ways. E.g. The recipient can use the Task.focus.identifier to positively match the Task.identifier of a Task it saw (or sent) previously. This allows a continuous chain of Tasks (additionally one would not need to retrieve Task instance referenced by the URL to establish the Task chain). A Task.focus.identifier can be used in a search to find a matching Task instance at a location (such as the Health Authority FHIR server) - this can be necessary as there is no guarantee that the Task.id will be the one originally requested. Of course, this requires the system+identifier in the identifier element to be globally unique (which should be true).